refactor(server): move document routes into package
